requirements.txt
altgraph==0.17.4
cryptography==41.0.7
Deprecated==1.2.14
et-xmlfile==1.1.0
img2pdf==0.5.1
lxml==4.9.4
openpyxl==3.1.2
oracledb==2.0.0
packaging==23.2
pandas==2.1.2
pefile==2023.2.7
pikepdf==8.10.1
Pillow==10.1.0
pycparser==2.21
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
PyQt6==6.6.0
PyQt6-Qt6==6.6.0
PyQt6-sip==13.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
wrapt==1.16.0
xlrd==2.0.1
